Text_generative_AI/requirements.txt
flask>=2.0.0
openai>=1.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
transformers>=4.20.0
torch>=1.9.0
accelerate>=0.20.0